% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Creates the Weekly Highlights Poll Slack Text.}
\usage{
run_app()
}
\description{
Creates the Weekly Highlights Poll Slack Text.
}
\details{
This app lets you pick up up to 10 links from the current
 R Weekly draft, to be voted upon by R Weekly editors.
  The app outputs the text to be pasted in the '#highlights' 
  Slack channel.
}
\examples{
\dontrun{
run_app()
}
}
